Folder Restructure - organization of stray and inappropriately named files
