+++ Languages/C++/Kedane'sAlgo.cpp
