+++ Languages/C++/numgen.cpp
