+++ Languages/Python/is_anagram.py
