+++ Languages/TypeScript/HelloWorldInTS.ts
